Fix CI error

+++ test/dummy/db/schema.rb
@@ -59,7 +59,7 @@ ActiveRecord::Schema.define(version: 2021_06_29_174736) do
   end
 
   create_table "active_storage_variant_records", force: :cascade do |t|
-    t.integer "blob_id", null: false
+    t.bigint "blob_id", null: false
     t.string "variation_digest", null: false
     t.index ["blob_id", "variation_digest"], name: "index_active_storage_variant_records_uniqueness", unique: true
   end
